`src/msac.rs`: Cleanup `as` casts.

+++ src/msac.rs
@@ -88,14 +88,14 @@ pub unsafe fn dav1d_msac_decode_bools(s: &mut MsacContext, n: libc::c_uint) -> l
 
 #[inline]
 pub unsafe fn dav1d_msac_decode_uniform(s: &mut MsacContext, n: libc::c_uint) -> libc::c_int {
-    if !(n > 0 as libc::c_uint) {
+    if !(n > 0) {
         unreachable!();
     }
     let l = ulog2(n) + 1;
     if !(l > 1) {
         unreachable!();
     }
-    let m = (((1 as libc::c_int) << l) as libc::c_uint).wrapping_sub(n);
+    let m = ((1 << l) as libc::c_uint).wrapping_sub(n);
     let v = dav1d_msac_decode_bools(s, (l - 1) as libc::c_uint);
     return (if v < m {
         v
@@ -113,13 +113,13 @@ unsafe extern "C" fn msac_init_x86(s: &mut MsacContext) {
     use crate::src::x86::cpu::DAV1D_X86_CPU_FLAG_SSE2;
 
     let flags = dav1d_get_cpu_flags();
-    if flags & DAV1D_X86_CPU_FLAG_SSE2 as libc::c_int as libc::c_uint != 0 {
+    if flags & DAV1D_X86_CPU_FLAG_SSE2 != 0 {
         s.symbol_adapt16 = Some(
             dav1d_msac_decode_symbol_adapt16_sse2
                 as unsafe extern "C" fn(*mut MsacContext, *mut uint16_t, size_t) -> libc::c_uint,
         );
     }
-    if flags & DAV1D_X86_CPU_FLAG_AVX2 as libc::c_int as libc::c_uint != 0 {
+    if flags & DAV1D_X86_CPU_FLAG_AVX2 != 0 {
         s.symbol_adapt16 = Some(
             dav1d_msac_decode_symbol_adapt16_avx2
                 as unsafe extern "C" fn(*mut MsacContext, *mut uint16_t, size_t) -> libc::c_uint,
@@ -142,7 +142,7 @@ unsafe extern "C" fn ctx_refill(s: &mut MsacContext) {
         let fresh1 = buf_pos;
         buf_pos = buf_pos.offset(1);
         dif ^= (*fresh1 as ec_win) << c;
-        c -= 8 as libc::c_int;
+        c -= 8;
     }
     s.dif = dif;
     s.cnt = EC_WIN_SIZE.wrapping_sub(c as usize).wrapping_sub(24) as libc::c_int;
@@ -151,8 +151,8 @@ unsafe extern "C" fn ctx_refill(s: &mut MsacContext) {
 
 #[inline]
 unsafe extern "C" fn ctx_norm(s: &mut MsacContext, dif: ec_win, rng: libc::c_uint) {
-    let d = 15 as libc::c_int ^ (31 as libc::c_int ^ clz(rng));
-    if !(rng <= 65535 as libc::c_uint) {
+    let d = 15 ^ (31 ^ clz(rng));
+    if !(rng <= 65535) {
         unreachable!();
     }
     s.cnt -= d;
@@ -169,15 +169,13 @@ unsafe fn dav1d_msac_decode_bool_equi_rust(s: &mut MsacContext) -> libc::c_uint 
     if !(dif >> EC_WIN_SIZE.wrapping_sub(16) < r as size_t) {
         unreachable!();
     }
-    let mut v = ((r >> 8) << 7).wrapping_add(4 as libc::c_int as libc::c_uint);
+    let mut v = ((r >> 8) << 7).wrapping_add(4);
     let vw = (v as ec_win) << EC_WIN_SIZE.wrapping_sub(16);
-    let ret = (dif >= vw) as libc::c_int as libc::c_uint;
+    let ret = (dif >= vw) as libc::c_uint;
     dif = dif.wrapping_sub((ret as size_t).wrapping_mul(vw)) as ec_win as ec_win;
-    v = v.wrapping_add(
-        ret.wrapping_mul(r.wrapping_sub((2 as libc::c_int as libc::c_uint).wrapping_mul(v))),
-    );
+    v = v.wrapping_add(ret.wrapping_mul(r.wrapping_sub((2 as libc::c_uint).wrapping_mul(v))));
     ctx_norm(s, dif, v);
-    return (ret == 0) as libc::c_int as libc::c_uint;
+    return (ret == 0) as libc::c_uint;
 }
 
 unsafe fn dav1d_msac_decode_bool_rust(s: &mut MsacContext, f: libc::c_uint) -> libc::c_uint {
@@ -186,16 +184,13 @@ unsafe fn dav1d_msac_decode_bool_rust(s: &mut MsacContext, f: libc::c_uint) -> l
     if !(dif >> EC_WIN_SIZE.wrapping_sub(16) < r as size_t) {
         unreachable!();
     }
-    let mut v =
-        ((r >> 8).wrapping_mul(f >> 6) >> 7 - 6).wrapping_add(4 as libc::c_int as libc::c_uint);
+    let mut v = ((r >> 8).wrapping_mul(f >> 6) >> 7 - 6).wrapping_add(4);
     let vw = (v as ec_win) << EC_WIN_SIZE.wrapping_sub(16);
-    let ret = (dif >= vw) as libc::c_int as libc::c_uint;
+    let ret = (dif >= vw) as libc::c_uint;
     dif = (dif).wrapping_sub((ret as size_t).wrapping_mul(vw)) as ec_win as ec_win;
-    v = v.wrapping_add(
-        ret.wrapping_mul(r.wrapping_sub((2 as libc::c_int as libc::c_uint).wrapping_mul(v))),
-    );
+    v = v.wrapping_add(ret.wrapping_mul(r.wrapping_sub((2 as libc::c_uint).wrapping_mul(v))));
     ctx_norm(s, dif, v);
-    return (ret == 0) as libc::c_int as libc::c_uint;
+    return (ret == 0) as libc::c_uint;
 }
 
 pub unsafe fn dav1d_msac_decode_subexp(
@@ -207,14 +202,12 @@ pub unsafe fn dav1d_msac_decode_subexp(
     if !(n >> k == 8) {
         unreachable!();
     }
-    let mut a = 0 as libc::c_int as libc::c_uint;
+    let mut a = 0;
     if dav1d_msac_decode_bool_equi(s) != 0 {
         if dav1d_msac_decode_bool_equi(s) != 0 {
-            k = k.wrapping_add(
-                (dav1d_msac_decode_bool_equi(s)).wrapping_add(1 as libc::c_int as libc::c_uint),
-            );
+            k = k.wrapping_add((dav1d_msac_decode_bool_equi(s)).wrapping_add(1));
         }
-        a = ((1 as libc::c_int) << k) as libc::c_uint;
+        a = 1 << k;
     }
     let v = (dav1d_msac_decode_bools(s, k)).wrapping_add(a);
     let ret = (if r#ref * 2 <= n {
@@ -238,17 +231,16 @@ unsafe fn dav1d_msac_decode_symbol_adapt_rust(
     if !(n_symbols <= 15) {
         unreachable!();
     }
-    if !(cdf[n_symbols] as libc::c_int <= 32) {
+    if !(cdf[n_symbols] <= 32) {
         unreachable!();
     }
     loop {
         val = val.wrapping_add(1);
         u = v;
-        v = r.wrapping_mul((cdf[val as usize] as libc::c_int >> 6) as libc::c_uint);
+        v = r.wrapping_mul((cdf[val as usize] >> 6) as libc::c_uint);
         v >>= 7 - 6;
         v = v.wrapping_add(
-            (4 as libc::c_int as libc::c_uint)
-                .wrapping_mul((n_symbols as libc::c_uint).wrapping_sub(val)),
+            (4 as libc::c_uint).wrapping_mul((n_symbols as libc::c_uint).wrapping_sub(val)),
         );
         if !(c < v) {
             break;
@@ -264,21 +256,19 @@ unsafe fn dav1d_msac_decode_symbol_adapt_rust(
     );
     if s.allow_update_cdf != 0 {
         let count = cdf[n_symbols] as libc::c_uint;
-        let rate = (4 as libc::c_int as libc::c_uint)
+        let rate = (4 as libc::c_uint)
             .wrapping_add(count >> 4)
             .wrapping_add((n_symbols > 2) as u32);
         let mut i = 0;
         while i < val {
-            cdf[i as usize] += (32768 - cdf[i as usize] as libc::c_int >> rate) as uint16_t;
+            cdf[i as usize] += 32768 - cdf[i as usize] >> rate;
             i = i.wrapping_add(1);
         }
         while (i as size_t) < n_symbols {
-            cdf[i as usize] -= (cdf[i as usize] as libc::c_int >> rate) as uint16_t;
+            cdf[i as usize] -= cdf[i as usize] >> rate;
             i = i.wrapping_add(1);
         }
-        cdf[n_symbols] = count
-            .wrapping_add((count < 32 as libc::c_uint) as libc::c_int as libc::c_uint)
-            as uint16_t;
+        cdf[n_symbols] = count.wrapping_add((count < 32) as libc::c_uint) as uint16_t;
     }
     return val;
 }
@@ -306,31 +296,28 @@ unsafe fn dav1d_msac_decode_bool_adapt_rust(
     let bit = dav1d_msac_decode_bool(s, cdf[0] as libc::c_uint);
     if s.allow_update_cdf != 0 {
         let count = cdf[1] as libc::c_uint;
-        let rate = (4 as libc::c_int as libc::c_uint).wrapping_add(count >> 4) as libc::c_int;
+        let rate = (4 as libc::c_uint).wrapping_add(count >> 4) as libc::c_int;
         if bit != 0 {
-            cdf[0] += (32768 as libc::c_int - cdf[0] as libc::c_int >> rate) as uint16_t;
+            cdf[0] += 32768 - cdf[0] >> rate;
         } else {
-            cdf[0] -= (cdf[0] as libc::c_int >> rate) as uint16_t;
+            cdf[0] -= cdf[0] >> rate;
         }
-        cdf[1] = count.wrapping_add((count < 32 as libc::c_uint) as libc::c_int as libc::c_uint)
-            as uint16_t;
+        cdf[1] = count.wrapping_add((count < 32) as libc::c_uint) as uint16_t;
     }
     return bit;
 }
 
 unsafe fn dav1d_msac_decode_hi_tok_rust(s: &mut MsacContext, cdf: &mut [u16; 4]) -> libc::c_uint {
-    let mut tok_br = dav1d_msac_decode_symbol_adapt4(s, cdf, 3 as libc::c_int as size_t);
-    let mut tok = (3 as libc::c_int as libc::c_uint).wrapping_add(tok_br);
-    if tok_br == 3 as libc::c_uint {
-        tok_br = dav1d_msac_decode_symbol_adapt4(s, cdf, 3 as libc::c_int as size_t);
-        tok = (6 as libc::c_int as libc::c_uint).wrapping_add(tok_br);
-        if tok_br == 3 as libc::c_uint {
-            tok_br = dav1d_msac_decode_symbol_adapt4(s, cdf, 3 as libc::c_int as size_t);
-            tok = (9 as libc::c_int as libc::c_uint).wrapping_add(tok_br);
-            if tok_br == 3 as libc::c_uint {
-                tok = (12 as libc::c_int as libc::c_uint).wrapping_add(
-                    dav1d_msac_decode_symbol_adapt4(s, cdf, 3 as libc::c_int as size_t),
-                );
+    let mut tok_br = dav1d_msac_decode_symbol_adapt4(s, cdf, 3);
+    let mut tok = (3 as libc::c_uint).wrapping_add(tok_br);
+    if tok_br == 3 {
+        tok_br = dav1d_msac_decode_symbol_adapt4(s, cdf, 3);
+        tok = (6 as libc::c_uint).wrapping_add(tok_br);
+        if tok_br == 3 {
+            tok_br = dav1d_msac_decode_symbol_adapt4(s, cdf, 3);
+            tok = (9 as libc::c_uint).wrapping_add(tok_br);
+            if tok_br == 3 {
+                tok = (12 as libc::c_uint).wrapping_add(dav1d_msac_decode_symbol_adapt4(s, cdf, 3));
             }
         }
     }
@@ -345,9 +332,9 @@ pub unsafe fn dav1d_msac_init(
 ) {
     s.buf_pos = data;
     s.buf_end = data.offset(sz as isize);
-    s.dif = ((1 as libc::c_int as ec_win) << EC_WIN_SIZE.wrapping_sub(1)).wrapping_sub(1);
-    s.rng = 0x8000 as libc::c_int as libc::c_uint;
-    s.cnt = -(15 as libc::c_int);
+    s.dif = ((1 as ec_win) << EC_WIN_SIZE.wrapping_sub(1)).wrapping_sub(1);
+    s.rng = 0x8000;
+    s.cnt = -15;
     s.allow_update_cdf = (disable_cdf_update_flag == 0) as libc::c_int;
     ctx_refill(s);
